Database.SQLite.Simple.FromRow

Description

The FromRow typeclass, for converting a row of results returned by a SQL query into a more useful Haskell representation.

Predefined instances are provided for tuples containing up to ten elements.

Documentation

class GFromRow f where

Generic derivation of FromRow.

Instantiating FromRow can in some cases be quite tedious. Luckily we can derive it generically in some cases where the type at hand has a Generic instance. The current implementation only works for a (n-ary) product types. So we would not be able to e.g. derive a FromRow instance for

data Bool = True | False

We can, however, derive a generic instance for the User type (see the example in FromRow).

Since: 0.4.18.1

Methods

gfromRow :: RowParser (f a)

class FromRow a where

A collection type that can be converted from a sequence of fields. Instances are provided for tuples up to 10 elements and lists of any length.

Note that instances can defined outside of sqlite-simple, which is often useful. For example, here's an instance for a user-defined pair:

data User = User { name :: String, fileQuota :: Int }

instance FromRow User where
    fromRow = User <$> field <*> field

The number of calls to field must match the number of fields returned in a single row of the query result. Otherwise, a ConversionFailed exception will be thrown.

Note the caveats associated with user-defined implementations of fromRow.

Generic implementation

Since version 0.4.18.1 it is possible in some cases to derive a generic implementation for FromRow. With a Generic instance for User, the example above could be written:

instance FromRow User where

With -XDeriveAnyClass -XDerivingStrategies the same can be written:

deriving anyclass instance FromRow User

For more details refer to GFromRow.

Minimal complete definition

Nothing

Methods

fromRow :: RowParser a

default fromRow :: Generic a => GFromRow (Rep a) => RowParser a

data RowParser a

field :: FromField a => RowParser a

fieldWith :: FieldParser a -> RowParser a

numFieldsRemaining :: RowParser Int