Database.SQLite.Simple.ToRow

Description

The ToRow typeclass, for rendering a collection of parameters to a SQL query.

Predefined instances are provided for tuples containing up to ten elements.

Documentation

class GToRow f where

Generic derivation of ToRow. For details about what can be derived refer to GFromRow.

Since: 0.4.18.1

Methods

gtoRow :: f a -> [SQLData]

class ToRow a where

A collection type that can be turned into a list of SQLData elements.

Since version 0.4.18.1 it is possible in some cases to derive a generic implementation for ToRow. Refer to the documentation for FromRow to see how this can be done.

Minimal complete definition

Nothing

Methods

toRow :: a -> [SQLData]

ToField a collection of values.

default toRow :: Generic a => GToRow (Rep a) => a -> [SQLData]