Database.SQLite.Simple.Internal

Description

Internal bits. This interface is less stable and can change at any time. In particular this means that while the rest of the sqlite-simple package endeavors to follow the package versioning policy, this module does not. Also, at the moment there are things in here that aren't particularly internal and are exported elsewhere; these will eventually disappear from this module.

Documentation

newtype Connection

Connection to an open database.

You can use connectionHandle to gain access to the underlying http://hackage.haskell.org/package/direct-sqlite connection. This may be useful if you need to access some direct-sqlite functionality that's not exposed in the sqlite-simple API. This should be a safe thing to do although mixing both APIs is discouraged.

Constructors

Connection

Fields

connectionHandle :: Database

data ColumnOutOfBounds

Constructors

ColumnOutOfBounds

Fields

errorColumnIndex :: Int

data Field

A Field represents metadata about a particular field

Constructors

Field

Fields

result :: SQLData

column :: !Int

newtype RowParseRO

Constructors

RowParseRO

Fields

nColumns :: Int

newtype RowParser a

Constructors

RP

Fields

unRP :: ReaderT RowParseRO (StateT (Int, [SQLData]) Ok) a

gettypename :: SQLData -> ByteString