Database.SQLite.Simple.Time

Description

Conversions to/from Haskell UTCTime and Day types for SQLite3. Offers better performance than direct use of time package's 'read'/'show' functionality.

The parsers are heavily adapted for the specific variant of ISO 8601 that SQLite uses, and the printers attempt to duplicate this syntax.

Documentation

module Database.SQLite.Simple.Time.Implementation