Database.SQLite.Simple.ToField

Description

The ToField typeclass, for rendering a parameter to an SQLite value to be bound as a SQL query parameter.

Documentation

class ToField a where

A type that may be used as a single parameter to a SQL query.

Methods

toField :: a -> SQLData

Prepare a value for substitution into a query string.